move promote_shape_t to xshape
