+++ test/test_xutils.cpp
@@ -12,7 +12,7 @@
 #include <tuple>
 #include <complex>
 #include "xtensor/xutils.hpp"
-#include "xtensor/xstorage.hpp" // has to be imported for promote_shape tests
+#include "xtensor/xshape.hpp"
 
 namespace xt
 {
